2.1.0: Added more monitor events. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2457 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/producer/layer.cpp
@@ -150,8 +150,8 @@ public:
 							<< monitor::event("frame")	% static_cast<int64_t>(frame_number_)
 														% static_cast<int64_t>((static_cast<int64_t>(foreground_->nb_frames()) - static_cast<int64_t>(auto_play_delta_ ? *auto_play_delta_ : 0)));
 
-			foreground_event_subject_ << monitor::event("type") % u8(foreground_->name());
-			background_event_subject_ << monitor::event("type") % u8(background_->name());
+			foreground_event_subject_ << monitor::event("type") % foreground_->name();
+			background_event_subject_ << monitor::event("type") % background_->name();
 				
 			return frame;
 		}
